refactor: Rename pick and onset submodules
Rename signal.pick -> signal.pickers and signal.onset -> signal.onsets.
Also renamed the pick.py and onset.py files (containing the base
classes) to base.py in the respective submodules.

Also:
- Update documentation to reflect the name changes
- Update the examples to reflect the name changes
- Update the setup.py file to correctly install the new submodules
- Update a few docstrings in signal/scan.py

All examples run and benchmarks pass tests.
