--- docs/submodules/quakemigrate.signal.onset.rst
@@ -1,23 +1,0 @@
-quakemigrate.signal.onset
-=========================
-
-.. automodule:: quakemigrate.signal.onset
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-quakemigrate.signal.onset.onset
--------------------------------
-
-.. automodule:: quakemigrate.signal.onset.onset
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-quakemigrate.signal.onset.staltaonset
--------------------------------------
-
-.. automodule:: quakemigrate.signal.onset.staltaonset
-    :members:
-    :undoc-members:
-    :show-inheritance:
