+++ docs/submodules/quakemigrate.signal.onsets.rst
@@ -1,0 +1,23 @@
+quakemigrate.signal.onsets
+==========================
+
+.. automodule:: quakemigrate.signal.onsets
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+quakemigrate.signal.onsets.base
+-------------------------------
+
+.. automodule:: quakemigrate.signal.onsets.base
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+quakemigrate.signal.onsets.stalta
+---------------------------------
+
+.. automodule:: quakemigrate.signal.onsets.stalta
+    :members:
+    :undoc-members:
+    :show-inheritance:
