--- docs/submodules/quakemigrate.signal.pick.rst
@@ -1,23 +1,0 @@
-quakemigrate.signal.pick
-========================
-
-.. automodule:: quakemigrate.signal.pick
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-quakemigrate.signal.pick.gaussianpicker
----------------------------------------
-
-.. automodule:: quakemigrate.signal.pick.gaussianpicker
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-quakemigrate.signal.pick.pick
------------------------------
-
-.. automodule:: quakemigrate.signal.pick.pick
-    :members:
-    :undoc-members:
-    :show-inheritance:
