+++ docs/submodules/quakemigrate.signal.pickers.rst
@@ -1,0 +1,23 @@
+quakemigrate.signal.pickers
+===========================
+
+.. automodule:: quakemigrate.signal.pickers
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+quakemigrate.signal.pickers.base
+--------------------------------
+
+.. automodule:: quakemigrate.signal.pickers.base
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+quakemigrate.signal.pickers.gaussian
+------------------------------------
+
+.. automodule:: quakemigrate.signal.pickers.gaussian
+    :members:
+    :undoc-members:
+    :show-inheritance:
