+++ examples/Icequake_Rutford/rutford_locate.py
@@ -6,8 +6,8 @@ Locate stage for the Rutford icequake example.
 
 from quakemigrate.io import Archive, read_lut, read_stations
 from quakemigrate.signal import QuakeScan
-from quakemigrate.signal.onset import STALTAOnset
-from quakemigrate.signal.pick import GaussianPicker
+from quakemigrate.signal.onsets import STALTAOnset
+from quakemigrate.signal.pickers import GaussianPicker
 
 # --- i/o paths ---
 station_file = "./inputs/rutford_stations.txt"
