+++ quakemigrate/signal/onsets/base.py
