+++ quakemigrate/signal/pickers/__init__.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 """
-The :mod:`quakemigrate.pick` module handles the picking of seismic phases. The
-default method makes the phase picks by fitting a 1-D Gaussian to the Onset
+The :mod:`quakemigrate.pickers` module handles the picking of seismic phases.
+The default method makes the phase picks by fitting a 1-D Gaussian to the Onset
 function.
 
 Feel free to contribute more phase picking methods!
@@ -14,5 +14,5 @@ Feel free to contribute more phase picking methods!
 
 """
 
-from .gaussianpicker import GaussianPicker  # NOQA
-from .pick import PhasePicker  # NOQA
+from .gaussian import GaussianPicker  # NOQA
+from .base import PhasePicker  # NOQA
