+++ quakemigrate/signal/pickers/gaussian.py
@@ -20,7 +20,7 @@ from scipy.optimize import curve_fit
 
 from quakemigrate.plot.phase_picks import pick_summary
 import quakemigrate.util as util
-from .pick import PhasePicker
+from .base import PhasePicker
 
 
 class GaussianPicker(PhasePicker):
