+++ setup.py
@@ -219,7 +219,8 @@ def setup_package():
         "packages": ["quakemigrate", "quakemigrate.core", "quakemigrate.io",
                      "quakemigrate.export", "quakemigrate.lut",
                      "quakemigrate.plot", "quakemigrate.signal",
-                     "quakemigrate.signal.onset", "quakemigrate.signal.pick",
+                     "quakemigrate.signal.onsets",
+                     "quakemigrate.signal.pickers",
                      "quakemigrate.signal.local_mag"],
         "ext_modules": get_extensions(),
         "package_data": get_package_data(),
